flux-path/cluster/suspend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
b2-utils==0.1.16
    # via -r requirements.in
backports-zoneinfo==0.2.1
    # via django
cachetools==5.3.0
    # via google-auth
certifi==2022.12.7
    # via
    #   kubernetes
    #   requests
charset-normalizer==3.1.0
    # via requests
django==4.2.3
    # via
    #   b2-utils
    #   django-model-utils
    #   djangorestframework
    #   model-bakery
django-model-utils==4.3.1
    # via b2-utils
djangorestframework==3.14.0
    # via b2-utils
faker==19.2.0
    # via b2-utils
google-auth==2.17.3
    # via kubernetes
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
kubernetes==26.1.0
    # via -r requirements.in
model-bakery==1.12.0
    # via b2-utils
oauthlib==3.2.2
    # via requests-oauthlib
pillow==10.0.0
    # via b2-utils
prometheus-client==0.16.0
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   faker
    #   kubernetes
pytz==2023.3
    # via djangorestframework
pyyaml==6.0
    # via kubernetes
requests==2.28.2
    # via
    #   kubernetes
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via kubernetes
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   kubernetes
    #   python-dateutil
sqlparse==0.4.4
    # via django
typing-extensions==4.7.1
    # via
    #   asgiref
    #   faker
urllib3==1.26.15
    # via
    #   kubernetes
    #   requests
validate-docbr==1.10.0
    # via b2-utils
websocket-client==1.5.1
    # via kubernetes

# The following packages are considered to be unsafe in a requirements file:
# setuptools
